Move the logic for validating a digest over to the digest_provider.py
